[5.9] sandbox: UInt8 is too small to hold a TCP or UDP port
Cherry pick of #6518

TCP and UDP port numbers are unsigned 16-bit integers, but
SandboxNetworkPermission uses UInt8s so trying to specify anything
but a small subset of low port numbers causes an error:

    permissions: [
        .allowNetworkConnections(
            scope: .all(ports: [443, 8080]),
            reason: "This command accesses the network"
        )
    ]

leads to

    error: integer literal '443' overflows when stored into 'Array<UInt8>.ArrayLiteralElement' (aka 'UInt8')
    error: integer literal '8080' overflows when stored into 'Array<UInt8>.ArrayLiteralElement' (aka 'UInt8')

This commit updates the type definitions and adds some higher-numbered
ports to the .allowNetworkConnections test cases.

+++ Sources/Commands/Utilities/DescribedPackage.swift
@@ -188,8 +188,8 @@ struct DescribedPackage: Encodable {
         struct Permission: Encodable {
             enum NetworkScope: Encodable {
                 case none
-                case local(ports: [UInt8])
-                case all(ports: [UInt8])
+                case local(ports: [Int])
+                case all(ports: [Int])
                 case docker
                 case unixDomainSocket
